pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thesiskit"
version = "0.1.0"
description = "Everything you need to ship academic research"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "ThesisKit Contributors"}
]
keywords = ["research", "academic", "paper", "automation", "llm", "citations"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "jinja2>=3.0",
    "python-slugify>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
]
arxiv = ["arxiv>=2.0"]
semantic_scholar = ["semanticscholar>=0.4"]

[project.scripts]
thesiskit = "thesiskit.cli:main"

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/thesiskit"
Documentation = "https://github.com/YOUR_USERNAME/thesiskit#readme"
Repository = "https://github.com/YOUR_USERNAME/thesiskit"
Issues = "https://github.com/YOUR_USERNAME/thesiskit/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
